mask_dashboard/requirements.txt
pandas
plotly-express
streamlit
mysql.connector
